requirements.txt
# Airflow
apache-airflow==2.7.1
apache-airflow-providers-snowflake==5.0.0

# Spark
pyspark==3.5.0
py4j==0.10.9.7

# Snowflake
snowflake-connector-python==3.2.0
snowflake-sqlalchemy==1.5.0

# Streamlit
streamlit==1.28.0
plotly==5.17.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# API
requests==2.31.0
python-dotenv==1.0.0

# Spark-Snowflake Connector (CRUCIAL)
net.snowflake:spark-snowflake_2.12:2.11.0-spark_3.3